LayoutTests/media/media-load-event-expected.txt

Test that media file is not reloaded when an element is inserted into the DOM.

RUN(mediaElement = document.createElement('audio'))
RUN(mediaElement.src = 'content/test.wav')
RUN(mediaElement.load())

EVENT(loadstart)
EVENT(durationchange)
EVENT(loadeddata)
EVENT(load)
EVENT(loadend)

RUN(document.getElementById('parent').appendChild(mediaElement))
RUN(mediaElement.play())

EVENT(play)
EVENT(playing)

END OF TEST
